+++ sources/Device/src/Interface/Handlers_d.cpp
@@ -170,11 +170,11 @@ static void SetFormWave(SimpleMessage *msg)
 }
 
 
-static void SetGeneratorParameter(SimpleMessage *msg, void(*func)(Chan::E, FloatValue))
+static void SetGeneratorParameter(SimpleMessage *msg, void(*func)(Chan::E, DoubleValue))
 {
     Chan ch(msg->TakeUINT8());
 
-    FloatValue value(0.0F);
+    DoubleValue value(0.0F);
     value.FromUINT64(msg->TakeUINT64());
 
     func(ch, value);
@@ -347,7 +347,7 @@ static void WriteRegister(SimpleMessage *msg)
 {
     Register reg(msg->TakeUINT8());
 
-    FloatValue value(0.0F);
+    DoubleValue value(0.0F);
     value.FromUINT64(msg->TakeUINT64());
 
     switch (reg)
