+++ sources/Panel/src/Generator/Generator_p.cpp
@@ -162,14 +162,14 @@ void PGenerator::SetParameter(ParameterChoice *param)
 void PGenerator::SetOffset(Chan::E ch, float offset)
 {
     // \todo √овнокод - запись параметра из двух мест
-    Message::Set::Offset(static_cast<uint8>(ch), FloatValue(offset).ToUINT64()).Transmit();
+    Message::Set::Offset(static_cast<uint8>(ch), DoubleValue(offset).ToUINT64()).Transmit();
 }
 
 
 void PGenerator::SetAmplitude(Chan::E ch, float amplitude)
 {
     // \todo √овнокод - запись параметра из двух мест
-    Message::Set::Amplitude(static_cast<uint8>(ch), FloatValue(amplitude).ToUINT64()).Transmit();
+    Message::Set::Amplitude(static_cast<uint8>(ch), DoubleValue(amplitude).ToUINT64()).Transmit();
 }
 
 
@@ -196,7 +196,7 @@ void PGenerator::SetParameter(ParameterValue *param)
         Command::RequestData
     };
 
-    FloatValue value = param->GetValue();
+    DoubleValue value = param->GetValue();
 
     Chan ch(param->GetForm()->GetWave()->GetChannel());
