+++ sources/Panel/src/Generator/MathSupport.h
@@ -1,7 +1,7 @@
 #pragma once
 
 
-struct FloatValue;
+struct DoubleValue;
 class Parameter;
 class ParameterValue;
 
@@ -28,7 +28,7 @@ struct Order
 
 struct MathFloatValue
 {
-    // ¬озвращает числовое представление FloatValue с numDigits значащими цифрами и зап€той.
+    // ¬озвращает числовое представление DoubleValue с numDigits значащими цифрами и зап€той.
     static pString GetIndicatedValue(const ParameterValue *param);
 
     // ¬озвращает позицию первого ненулевого символа "1" - дес€тки (1e1), "0" - единицы (1e0), "-1" - дес€тые (1e-1), "-2" - сотые (1e-2)
@@ -36,23 +36,23 @@ struct MathFloatValue
 
     // ¬озвращает цифру в позиции position. “очка находитс€ соответственно order. One - после единиц, Kilo - после тыс€ч и так далее.
     // Order::Count - значенине по умолчанию - зап€та€ в позиции относительно размерности числового значени€
-    static int GetDigit(const FloatValue &value, int position, Order::E order = Order::Count);
+    static int GetDigit(const DoubleValue &value, int position, Order::E order = Order::Count);
 
     // ¬озвращает символ в позиции position. «нак не учитываетс€. “очка находитс€ соответственно order. One - после единиц, Kilo - после тыс€ч и так далее.
     // Order::Count - значенине по умолчанию - зап€та€ в позиции относительно размерности числового значени€
-    static char GetChar(const FloatValue &value, int postition, Order::E order = Order::Count);
+    static char GetChar(const DoubleValue &value, int postition, Order::E order = Order::Count);
 
     // ¬озвращает:
-    // Order::Mega  - FloatValue::Integer() >= 1e6
-    // Order::Kilo  - FloatValue::Integer() >= 1e3
-    // Order::One   - FloatValue::Integer() >= 0
-    // Order::Milli - FloatValue::Integer() >= 1e-3
-    // Order::Micro - FloatValue::Integer() >= 1e-6
-    static Order::E GetOrder(FloatValue value);
+    // Order::Mega  - DoubleValue::Integer() >= 1e6
+    // Order::Kilo  - DoubleValue::Integer() >= 1e3
+    // Order::One   - DoubleValue::Integer() >= 0
+    // Order::Milli - DoubleValue::Integer() >= 1e-3
+    // Order::Micro - DoubleValue::Integer() >= 1e-6
+    static Order::E GetOrder(DoubleValue value);
 
 private:
 
-    static void CorrectValueOnOrder(FloatValue *value, Order::E order);
+    static void CorrectValueOnOrder(DoubleValue *value, Order::E order);
 };
